requirements.txt
altgraph==0.16.1
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
future==0.17.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.1
idna==2.8
macholib==1.11
oauthlib==3.0.1
pefile==2018.8.8
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyInstaller==3.4
pywin32==224
pywin32-ctypes==0.2.0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9
uritemplate==3.0.0
urllib3==1.24.1
